backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
langchain
langchain-groq
langchain-community
chromadb
supabase
python-multipart
websockets
httpx
tavily-python